+++ projects/vr-lib/src/public-api.ts
@@ -5,3 +5,5 @@
 export { VrKitHeaderComponent } from './lib/components/header/header.component';
 export { VrKitPageHeaderComponent } from './lib/components/page-header/page-header.component';
 export { VrKitBreadcrumbComponent } from './lib/components/breadcrumb/breadcrumb.component';
+export { VrKitArticleComponent } from './lib/components/article/article.component';
+export { VrKitStepperComponent } from './lib/components/stepper/stepper.component';